# api-contracts/openapi/agent_api.yaml
openapi: 3.0.3
info:
  title: SMT Python Agent REST API
  description: |
    SMT 智能运维平台 Python 智能体对外 RESTful 接口契约。

    覆盖两个 Python Agent：
    - agent-knowledge（端口 8004）：知识助手 RAG 问答，支持文档上传、向量检索问答、文档列表与删除。
    - agent-maintenance（端口 8002）：设备运维 Agent，支持健康评分、故障诊断、趋势预测、案例沉淀。

    外部访问路径前缀 `/api/agent/**` 经 `smt-gateway` 网关路由到对应 Python 服务，
    网关 `StripPrefix=2` 后 Python 服务收到 `/knowledge/**` 与 `/maintenance/**`。
    响应体为业务对象直返（不使用 Result 包装），字段命名采用 snake_case 与 Python Pydantic 默认对齐。
  version: 1.0.0
  contact:
    name: smt-agent-platform
    url: https://github.com/north30/smt-agent-platform
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8080
    description: 本地开发环境（经 smt-gateway 网关）
...
paths:
  /api/agent/knowledge/upload:
    post:
      tags:
        - 知识助手
      summary: 上传知识文档
      description: |
        以 multipart/form-data 方式上传文档（PDF/Word/TXT/Markdown 等），服务端完成解析、分块、向量化后写入 Milvus。
        返回新文档的 ID、名称与分块数量。
      operationId: uploadKnowledgeDocument
      requestBody:
        required: true
        description: 待上传的文档文件
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/KnowledgeUploadRequest'
      responses:
        '200':
          description: 上传并入库成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeUploadResponse'
              examples:
                success:
                  summary: 上传成功
                  value:
                    doc_id: doc-2026-0001
                    doc_name: SMT印刷机维护手册.pdf
                    chunk_count: 42
        '400':
          description: 参数错误（文件为空、格式不支持、文件过大等）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emptyFile:
                  summary: 文件为空
                  value:
                    error: INVALID_FILE
                    message: 上传文件不能为空
  /api/agent/knowledge/ask:
    post:
      tags:
        - 知识助手
      summary: 知识库问答
      description: |
        基于已入库文档构建的知识库进行 RAG 检索增强问答：先在 Milvus 中检索与问题最相关的文本块，
        再交由大模型生成答案，并返回命中的来源片段（含 doc_id、chunk_id、相似度得分、内容片段）。
      operationId: askKnowledge
      requestBody:
        required: true
        description: 问答请求体
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeAskRequest'
            examples:
              printer:
                summary: 印刷机保养问题
                value:
                  question: 印刷机钢网清洁频率应该是多少？
      responses:
        '200':
          description: 问答成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeAskResponse'
              examples:
                answer:
                  summary: 问答示例
                  value:
                    answer: 建议每班次生产结束后清洁钢网一次...
                    sources:
                      - doc_id: doc-2026-0001
                        chunk_id: chunk-003
                        score: 0.92
                        snippet: 钢网清洁：每班次生产结束后使用专用清洗剂...
        '400':
          description: 参数错误（question 为空或超长）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emptyQuestion:
                  summary: 问题为空
                  value:
                    error: INVALID_QUESTION
                    message: 问题不能为空
...
  /api/agent/knowledge/documents/{doc_id}:
    delete:
      tags:
        - 知识助手
      summary: 删除知识库文档
      description: |
        根据文档 ID 删除知识库中的文档及其全部分块向量，返回是否删除成功与实际删除的分块数量。
      operationId: deleteKnowledgeDocument
      parameters:
        - $ref: '#/components/parameters/DocIdPathParam'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeDeleteResponse'
              examples:
                success:
                  summary: 删除成功
                  value:
                    success: true
                    deleted_chunks: 42
        '400':
          description: 参数错误（doc_id 格式不合法）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidId:
                  summary: 文档 ID 不合法
                  value:
                    error: INVALID_DOC_ID
                    message: 文档 ID 格式不合法
...
  /api/agent/maintenance/diagnose:
    post:
      tags:
        - 设备运维
      summary: 故障诊断
      description: |
        根据设备 ID 与故障现象描述，结合历史案例库与大模型进行根因分析，
        返回可能的根因列表、维修建议列表以及相似历史案例（含相似度得分）。
        诊断过程可能调用 smt-device-service 获取设备上下文，Java 服务不可达时返回 503。
      operationId: diagnoseDevice
      requestBody:
        required: true
        description: 故障诊断请求体
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceDiagnoseRequest'
            examples:
              vibration:
                summary: 振动异常诊断
                value:
                  device_id: 1001
                  symptom: 贴片机主轴振动幅值持续偏高
      responses:
        '200':
          description: 诊断成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceDiagnoseResponse'
              examples:
                diagnose:
                  summary: 诊断结果示例
                  value:
                    root_causes:
                      - 主轴轴承磨损
                      - 安装基础松动
                    repair_suggestions:
                      - 检查并更换主轴轴承
                      - 重新校准设备安装水平
                    similar_cases:
                      - case_id: case-0057
                        symptom: 主轴振动偏高伴随异响
                        root_cause: 主轴轴承磨损
                        solution: 更换主轴轴承并复测振动
                        score: 0.87
        '400':
          description: 参数错误（device_id 缺失或 symptom 为空）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emptySymptom:
                  summary: 故障现象为空
                  value:
                    error: INVALID_SYMPTOM
                    message: 故障现象不能为空
        '503':
          description: 依赖的 Java 服务（smt-device-service）不可达
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                upstream:
                  summary: Java 服务不可达
                  value:
                    error: DEVICE_SERVICE_UNAVAILABLE
                    message: 设备服务暂不可用
  /api/agent/maintenance/predict/{device_id}:
    get:
      tags:
        - 设备运维
      summary: 设备趋势预测
      description: |
        基于设备历史采集数据预测关键指标的未来走势，输出阈值越限告警、预测到达阈值的时间以及运维建议。
        当历史数据不足以支撑预测时 `data_sufficient=false` 并返回空预测。
        依赖 smt-device-service 提供历史数据，Java 服务不可达时返回 503。
      operationId: predictDeviceTrend
      parameters:
        - $ref: '#/components/parameters/DeviceIdPathParam'
      responses:
        '200':
          description: 预测成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenancePredictResponse'
              examples:
                predict:
                  summary: 预测结果示例
                  value:
                    trend: 上升
                    threshold_alerts:
                      - datapoint_code: TEMP-01
                        current_value: 78.5
                        threshold: 85.0
                        severity: WARNING
                    forecast:
                      datapoint_code: TEMP-01
                      predicted_value: 86.2
                      hours_to_threshold: 12
                    recommendation: 建议在 12 小时内检查散热系统并降低负载。
                    data_sufficient: true
        '400':
          description: 参数错误（device_id 不合法）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidId:
                  summary: 设备 ID 不合法
                  value:
                    error: INVALID_DEVICE_ID
                    message: 设备 ID 不合法
        '503':
          description: 依赖的 Java 服务（smt-device-service）不可达
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                upstream:
                  summary: Java 服务不可达
                  value:
                    error: DEVICE_SERVICE_UNAVAILABLE
                    message: 设备服务暂不可用
  /api/agent/maintenance/cases:
    post:
      tags:
        - 设备运维
      summary: 新增运维案例
      description: |
        将一次运维处置经验沉淀为案例写入案例库（Milvus），供后续诊断检索复用。
        请求体包含设备类型、故障现象、根因与解决方案，返回新建案例 ID。
      operationId: createMaintenanceCase
      requestBody:
        required: true
        description: 运维案例创建请求体
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceCaseCreateRequest'
            examples:
              case:
                summary: 案例创建示例
                value:
                  device_type: MOUNTER
                  symptom: 主轴振动偏高伴随异响
                  root_cause: 主轴轴承磨损
                  solution: 更换主轴轴承并复测振动
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceCaseCreateResponse'
              examples:
                success:
                  summary: 创建成功
                  value:
                    case_id: case-0058
        '400':
          description: 参数错误（必填字段缺失）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingField:
                  summary: 必填字段缺失
                  value:
                    error: MISSING_CASE_REQUIRED_FIELDS
                    message: 设备类型、故障现象、根因、解决方案均为必填
...
components:
  parameters:
    DocIdPathParam:
      name: doc_id
      in: path
      description: 知识库文档 ID
      required: true
      schema:
        type: string
        example: doc-2026-0001
    DeviceIdPathParam:
      name: device_id
      in: path
      description: 设备 ID
      required: true
      schema:
        type: integer
        format: int64
        example: 1001
...
  schemas:
    ErrorResponse:
      type: object
      description: 错误响应
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: 错误码（大写蛇形，如 INVALID_FILE、DEVICE_NOT_FOUND）
          example: INVALID_FILE
        message:
          type: string
          description: 错误描述
          example: 上传文件不能为空
    KnowledgeUploadRequest:
      type: object
      description: 知识文档上传请求体（multipart/form-data）
      required:
        - file
      properties:
        file:
          type: string
          format: binary
          description: 待上传的文档文件（支持 PDF/Word/TXT/Markdown）
    KnowledgeUploadResponse:
      type: object
      description: 知识文档上传响应
      required:
        - doc_id
        - doc_name
        - chunk_count
      properties:
        doc_id:
          type: string
          description: 文档 ID
          example: doc-2026-0001
        doc_name:
          type: string
          description: 文档名称
          example: SMT印刷机维护手册.pdf
        chunk_count:
          type: integer
          description: 分块数量
          example: 42
    KnowledgeAskRequest:
      type: object
      description: 知识库问答请求体
      required:
        - question
      properties:
        question:
          type: string
          description: 用户问题
          example: 印刷机钢网清洁频率应该是多少？
    KnowledgeSource:
      type: object
      description: 知识库问答来源片段
      required:
        - doc_id
        - chunk_id
        - score
        - snippet
      properties:
        doc_id:
          type: string
          description: 来源文档 ID
          example: doc-2026-0001
        chunk_id:
          type: string
          description: 来源分块 ID
          example: chunk-003
        score:
          type: number
          format: float
          description: 相似度得分（0-1，越大越相关）
          minimum: 0
          maximum: 1
          example: 0.92
        snippet:
          type: string
          description: 命中文本片段
          example: 钢网清洁：每班次生产结束后使用专用清洗剂...
    KnowledgeAskResponse:
      type: object
      description: 知识库问答响应
      required:
        - answer
        - sources
      properties:
        answer:
          type: string
          description: 大模型生成的答案
          example: 建议每班次生产结束后清洁钢网一次...
        sources:
          type: array
          description: 命中的来源片段列表
          items:
            $ref: '#/components/schemas/KnowledgeSource'
...
    KnowledgeDeleteResponse:
      type: object
      description: 知识库文档删除响应
      required:
        - success
        - deleted_chunks
      properties:
        success:
          type: boolean
          description: 是否删除成功
          example: true
        deleted_chunks:
          type: integer
          description: 实际删除的分块数量
          example: 42
...
    MaintenanceDiagnoseRequest:
      type: object
      description: 故障诊断请求体
      required:
        - device_id
        - symptom
      properties:
        device_id:
          type: integer
          format: int64
          description: 设备 ID
          example: 1001
        symptom:
          type: string
          description: 故障现象描述
          example: 贴片机主轴振动幅值持续偏高
    MaintenanceSimilarCase:
      type: object
      description: 相似历史案例
      required:
        - case_id
        - symptom
        - root_cause
        - solution
        - score
      properties:
        case_id:
          type: string
          description: 案例 ID
          example: case-0057
        symptom:
          type: string
          description: 故障现象
          example: 主轴振动偏高伴随异响
        root_cause:
          type: string
          description: 根因
          example: 主轴轴承磨损
        solution:
          type: string
          description: 解决方案
          example: 更换主轴轴承并复测振动
        score:
          type: number
          format: float
          description: 相似度得分（0-1）
          minimum: 0
          maximum: 1
          example: 0.87
    MaintenanceDiagnoseResponse:
      type: object
      description: 故障诊断响应
      required:
        - root_causes
        - repair_suggestions
        - similar_cases
      properties:
        root_causes:
          type: array
          description: 可能的根因列表
          items:
            type: string
          example:
            - 主轴轴承磨损
            - 安装基础松动
        repair_suggestions:
          type: array
          description: 维修建议列表
          items:
            type: string
          example:
            - 检查并更换主轴轴承
            - 重新校准设备安装水平
        similar_cases:
          type: array
          description: 相似历史案例列表
          items:
            $ref: '#/components/schemas/MaintenanceSimilarCase'
    MaintenanceThresholdAlert:
      type: object
      description: 阈值越限告警
      required:
        - datapoint_code
        - current_value
        - threshold
        - severity
      properties:
        datapoint_code:
          type: string
          description: 采集点编码
          example: TEMP-01
        current_value:
          type: number
          format: float
          description: 当前值
          example: 78.5
        threshold:
          type: number
          format: float
          description: 阈值
          example: 85.0
        severity:
          type: string
          description: 严重等级
          enum:
            - INFO
            - WARNING
            - CRITICAL
          example: WARNING
    MaintenanceForecast:
      type: object
      description: 指标预测结果
      required:
        - datapoint_code
        - predicted_value
        - hours_to_threshold
      properties:
        datapoint_code:
          type: string
          description: 采集点编码
          example: TEMP-01
        predicted_value:
          type: number
          format: float
          description: 预测值
          example: 86.2
        hours_to_threshold:
          type: integer
          description: 预计到达阈值的小时数
          example: 12
    MaintenancePredictResponse:
      type: object
      description: 设备趋势预测响应
      required:
        - trend
        - threshold_alerts
        - forecast
        - recommendation
        - data_sufficient
      properties:
        trend:
          type: string
          description: 趋势方向（上升/下降/平稳）
          example: 上升
        threshold_alerts:
          type: array
          description: 阈值越限告警列表
          items:
            $ref: '#/components/schemas/MaintenanceThresholdAlert'
        forecast:
          $ref: '#/components/schemas/MaintenanceForecast'
        recommendation:
          type: string
          description: 运维建议（大模型生成）
          example: 建议在 12 小时内检查散热系统并降低负载。
        data_sufficient:
          type: boolean
          description: 历史数据是否足以支撑预测
          example: true
    MaintenanceCaseCreateRequest:
      type: object
      description: 运维案例创建请求体
      required:
        - device_type
        - symptom
        - root_cause
        - solution
      properties:
        device_type:
          type: string
          description: 设备类型
          example: MOUNTER
        symptom:
          type: string
          description: 故障现象
          example: 主轴振动偏高伴随异响
        root_cause:
          type: string
          description: 根因
          example: 主轴轴承磨损
        solution:
          type: string
          description: 解决方案
          example: 更换主轴轴承并复测振动
    MaintenanceCaseCreateResponse:
      type: object
      description: 运维案例创建响应
      required:
        - case_id
      properties:
        case_id:
          type: string
          description: 新建案例 ID
          example: case-0058
...